On second thought... let's use a variable `this` parameter in `onChanging`
And the holder itself is rarely useful, so drop that parameter.

For React components, `this` should be the component. For `holdProps` it
is the model, by default, but can also be customized. For `holdValue`,
the default `this` parameter is the `ValueHolder`, but it can be
customized also. For `holdStates`, `this` will always be the component.